Created HomeCommand.java
