+++ src/main/java/dansplugins/sethomesystem/CommandInterpreter.java
@@ -1,5 +1,6 @@
 package dansplugins.sethomesystem;
 
+import dansplugins.sethomesystem.commands.HomeCommand;
 import dansplugins.sethomesystem.commands.SetHomeCommand;
 import dansplugins.sethomesystem.data.PersistentData;
 import dansplugins.sethomesystem.managers.StorageManager;
@@ -36,42 +37,8 @@ public class CommandInterpreter {
 
         // home command
         if (label.equalsIgnoreCase("home")) {
-            if (sender instanceof Player) {
-                Player player = (Player) sender;
-                for (HomeRecord record : PersistentData.getInstance().getHomeRecords()) {
-                    if (record.getPlayerName().equalsIgnoreCase(player.getName())) {
-                        if (record.getHomeLocation() != null) {
-                            player.sendMessage(ChatColor.GREEN + "Teleporting in 3 seconds...");
-                            int seconds = 3;
-
-                            Location initialLocation = player.getLocation();
-
-                            getServer().getScheduler().runTaskLater(MedievalSetHome.getInstance(), new Runnable() {
-                                @Override
-                                public void run() {
-                                    if (initialLocation.getX() == player.getLocation().getX() &&
-                                            initialLocation.getY() == player.getLocation().getY() &&
-                                            initialLocation.getZ() == player.getLocation().getZ()) {
-
-                                        // teleport the player
-                                        player.teleport(record.getHomeLocation());
-
-                                    }
-                                    else {
-                                        player.sendMessage(ChatColor.RED + "Movement Detected. Teleport cancelled.");
-                                    }
-
-                                }
-                            }, seconds * 20);
-                        }
-                        else {
-                            player.sendMessage(ChatColor.RED + "Home not set!");
-                        }
-
-                        return true;
-                    }
-                }
-            }
+            HomeCommand command = new HomeCommand();
+            return command.teleportPlayerHome(sender);
         }
 
         // msh command
